LayoutTests/editing/pasteboard/paste-table-with-unrendered-text-nodes-expected.txt


Initial state:
| "line 1"
| <br>
| <#selection-caret>
| <br>

After pasting line 2 (table A):
| "line 1"
| <br>
| <table>
|   border="1"
|   "\n"
|   <tbody>
|     "\n"
|     <tr>
|       "\n"
|       <td>
|         "line 2 table A"
|       "\n"
|     "\n"
|   "\n"
| <#selection-caret>

After typing ling 3:
| "line 1"
| <br>
| <table>
|   border="1"
|   "\n"
|   <tbody>
|     "\n"
|     <tr>
|       "\n"
|       <td>
|         "line 2 table A"
|       "\n"
|     "\n"
|   "\n"
| "line 3<#selection-caret>"

After pasting ling 4 (table B):
| "line 1"
| <br>
| <table>
|   border="1"
|   "\n"
|   <tbody>
|     "\n"
|     <tr>
|       "\n"
|       <td>
|         "line 2 table A"
|       "\n"
|     "\n"
|   "\n"
| "line 3"
| <table>
|   border="1"
|   "\n"
|   <tbody>
|     "\n"
|     <tr>
|       "\n"
|       <td>
|         "line 4"
|       "\n"
|       <td>
|         "table B"
|       "\n"
|     "\n"
|   "\n"
| <#selection-caret>

After typing ling 5:
| "line 1"
| <br>
| <table>
|   border="1"
|   "\n"
|   <tbody>
|     "\n"
|     <tr>
|       "\n"
|       <td>
|         "line 2 table A"
|       "\n"
|     "\n"
|   "\n"
| "line 3"
| <table>
|   border="1"
|   "\n"
|   <tbody>
|     "\n"
|     <tr>
|       "\n"
|       <td>
|         "line 4"
|       "\n"
|       <td>
|         "table B"
|       "\n"
|     "\n"
|   "\n"
| "line 5<#selection-caret>"
